requirements.txt
torch
numpy
tensorflow
tensorboard
tensorboardx
progressbar2

